API: simpler and more consistent path arguments
Before this change, the template notebook was a filename with no extension.
A second argument specified the folder. Now, the template notebook is
specified with a single argument containing its full path.
The output path argument defaults to the template notebook folder.
Path arguments can be strings or pathlib.Path objects.